feat: refactor component
