+++ src/pages/Game/components/Questions/styles.css
@@ -1,7 +1,7 @@
-/* .answer-btn-correct {
+.answers.visible .answer-btn-correct {
   border: 3px solid rgb(6, 240, 15);
 }
 
-.answer-btn-inc {
+.answers.visible .answer-btn-inc {
   border: 3px solid rgb(255, 0, 0);
-} */
+}
